ThresholdEvent has to be an interface
