+++ lib/src/Uncertainty/Algorithm/Simulation/EventSimulation.cxx
@@ -29,6 +29,7 @@
 #include "openturns/Less.hxx"
 #include "openturns/Uniform.hxx"
 #include "openturns/IdentityFunction.hxx"
+#include "openturns/CompositeRandomVector.hxx"
 
 BEGIN_NAMESPACE_OPENTURNS
 
@@ -43,7 +44,7 @@ static const Factory<EventSimulation> Factory_EventSimulation;
 /** For save/load mechanism */
 EventSimulation::EventSimulation(const Bool verbose, const HistoryStrategy & convergenceStrategy)
   : SimulationAlgorithm()
-  , event_(new ThresholdEvent(CompositeRandomVector(IdentityFunction(1), RandomVector(Uniform())), Less(), 0.0))
+  , event_(ThresholdEvent(CompositeRandomVector(IdentityFunction(1), RandomVector(Uniform())), Less(), 0.0))
   , result_()
 {
   setVerbose(verbose);
