+++ lib/src/Uncertainty/Model/CMakeLists.txt
@@ -22,6 +22,7 @@ ot_add_source_file (KrigingRandomVector.cxx)
 ot_add_source_file (UsualRandomVector.cxx)
 ot_add_source_file (DomainEvent.cxx)
 ot_add_source_file (ThresholdEvent.cxx)
+ot_add_source_file (ThresholdEventImplementation.cxx)
 ot_add_source_file (ProcessEvent.cxx)
 ot_add_source_file (IntersectionEvent.cxx)
 ot_add_source_file (StandardEvent.cxx)
@@ -53,6 +54,7 @@ ot_install_header_file (UnionEvent.hxx)
 ot_install_header_file (Distribution.hxx)
 ot_install_header_file (DomainEvent.hxx)
 ot_install_header_file (ThresholdEvent.hxx)
+ot_install_header_file (ThresholdEventImplementation.hxx)
 ot_install_header_file (ProcessEvent.hxx)
 ot_install_header_file (IntersectionEvent.hxx)
 ot_install_header_file (ArchimedeanCopula.hxx)
