+++ lib/test/t_DirectionalSampling_std.expout
@@ -1,2 +1,2 @@
-DirectionalSampling=class=DirectionalSampling rootStrategy=class=RootStrategy implementation=class=SafeAndSlow derived from class=RootStrategyImplementation solver=class=Solver implementation=class=Brent derived from class=SolverImplementation absoluteError=1e-05 relativeError=1e-05 residualError=1e-08 maximumFunctionEvaluation=100 usedFunctionEvaluation=0 maximumDistance=8 stepSize=1 samplingStrategy=class=SamplingStrategy implementation=class=RandomDirection derived from class=SamplingStrategyImplementation dimension=4 derived from class=EventSimulation event=class=RandomVector implementation=class=ThresholdEvent antecedent=class=CompositeRandomVector function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[E,F,L,I,y0] evaluationImplementation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] gradientImplementation=class=SymbolicGradient name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] hessianImplementation=class=SymbolicHessian name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] antecedent=class=UsualRandomVector distribution=class=Normal name=Normal dimension=4 mean=class=Point name=Unnamed dimension=4 values=[50,1,10,5] sigma=class=Point name=Unnamed dimension=4 values=[1,1,1,1] correlationMatrix=class=CorrelationMatrix dimension=4 implementation=class=MatrixImplementation name=Unnamed rows=4 columns=4 values=[1,0,0,0,0,1,0,0,0,0,1,0,0,0,0,1] operator=class=Less name=Unnamed threshold=-3 maximumOuterSampling=250 maximumCoefficientOfVariation=0.1 maximumStandardDeviation=0 blockSize=4
+DirectionalSampling=class=DirectionalSampling rootStrategy=class=RootStrategy implementation=class=SafeAndSlow derived from class=RootStrategyImplementation solver=class=Solver implementation=class=Brent derived from class=SolverImplementation absoluteError=1e-05 relativeError=1e-05 residualError=1e-08 maximumFunctionEvaluation=100 usedFunctionEvaluation=0 maximumDistance=8 stepSize=1 samplingStrategy=class=SamplingStrategy implementation=class=RandomDirection derived from class=SamplingStrategyImplementation dimension=4 derived from class=EventSimulation event=class=RandomVector implementation=class=ThresholdEventImplementation antecedent=class=CompositeRandomVector function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[E,F,L,I,y0] evaluationImplementation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] gradientImplementation=class=SymbolicGradient name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] hessianImplementation=class=SymbolicHessian name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] antecedent=class=UsualRandomVector distribution=class=Normal name=Normal dimension=4 mean=class=Point name=Unnamed dimension=4 values=[50,1,10,5] sigma=class=Point name=Unnamed dimension=4 values=[1,1,1,1] correlationMatrix=class=CorrelationMatrix dimension=4 implementation=class=MatrixImplementation name=Unnamed rows=4 columns=4 values=[1,0,0,0,0,1,0,0,0,0,1,0,0,0,0,1] operator=class=Less name=Unnamed threshold=-3 maximumOuterSampling=250 maximumCoefficientOfVariation=0.1 maximumStandardDeviation=0 blockSize=4
 DirectionalSampling result=probabilityEstimate=1.748122e-01 varianceEstimate=2.978217e-04 standard deviation=1.73e-02 coefficient of variation=9.87e-02 confidenceLength(0.95)=6.76e-02 outerSampling=30 blockSize=4
