+++ lib/test/t_Event_std.expout
@@ -3,7 +3,7 @@ vect=class=ConstantRandomVector description=[x0,x1,x2,x3] point=class=Point name
 vect dimension=4
 vect realization (first )=class=Point name=Unnamed dimension=4 values=[101,202,303,404]
 vect sample =class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=5 dimension=4 description=[x0,x1,x2,x3] data=[[101,202,303,404],[101,202,303,404],[101,202,303,404],[101,202,303,404],[101,202,303,404]]
-myEvent=class=ThresholdEvent antecedent=class=CompositeRandomVector function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[E,F,L,I,y0] evaluationImplementation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] gradientImplementation=class=SymbolicGradient name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] hessianImplementation=class=SymbolicHessian name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] antecedent=class=ConstantRandomVector description=[x0,x1,x2,x3] point=class=Point name=Unnamed dimension=4 values=[101,202,303,404] operator=class=Less name=Unnamed threshold=50
+myEvent=class=ThresholdEventImplementation antecedent=class=CompositeRandomVector function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[E,F,L,I,y0] evaluationImplementation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] gradientImplementation=class=SymbolicGradient name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] hessianImplementation=class=SymbolicHessian name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] antecedent=class=ConstantRandomVector description=[x0,x1,x2,x3] point=class=Point name=Unnamed dimension=4 values=[101,202,303,404] operator=class=Less name=Unnamed threshold=50
 myEvent realization=class=Point name=Unnamed dimension=1 values=[1]
 myEvent antecedent realization=class=Point name=Unnamed dimension=4 values=[101,202,303,404]
-myEvent sample=class=Sample name=ThresholdEvent sample implementation=class=SampleImplementation name=ThresholdEvent sample size=10 dimension=1 description=[y0] data=[[1],[1],[1],[1],[1],[1],[1],[1],[1],[1]]
+myEvent sample=class=Sample name=ThresholdEventImplementation sample implementation=class=SampleImplementation name=ThresholdEventImplementation sample size=10 dimension=1 description=[y0] data=[[1],[1],[1],[1],[1],[1],[1],[1],[1],[1]]
