+++ lib/test/t_FORM_SQP.expout
@@ -1,4 +1,4 @@
-FORM=class=FORM class=Analytical nearestPointAlgorithm=class=OptimizationAlgorithm implementation=class=SQP class=OptimizationAlgorithmImplementation problem=class=OptimizationProblem implementation=class=NearestPointProblem level function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[x1,x2,x3,x4,y0] evaluationImplementation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[x1,x2,x3,x4] outputVariablesNames=[y0] formulas=[x1+2*x2-3*x3+4*x4] gradientImplementation=class=SymbolicGradient name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[x1,x2,x3,x4] outputVariablesNames=[y0] formulas=[x1+2*x2-3*x3+4*x4] hessianImplementation=class=SymbolicHessian name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[x1,x2,x3,x4] outputVariablesNames=[y0] formulas=[x1+2*x2-3*x3+4*x4] level value=3 dimension=4 startingPoint=class=Point name=Unnamed dimension=4 values=[1,1,1,1] maximumIterationNumber=100 maximumEvaluationNumber=100000 maximumAbsoluteError=1e-10 maximumRelativeError=1e-10 maximumResidualError=1e-10 maximumConstraintError=1e-10 verbose=false tau=0.5 omega=0.0001 smooth=1.2 event=class=RandomVector implementation=class=ThresholdEvent antecedent=class=CompositeRandomVector function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[E,F,L,I,y0] evaluationImplementation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] gradientImplementation=class=SymbolicGradient name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] hessianImplementation=class=SymbolicHessian name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] antecedent=class=UsualRandomVector distribution=class=Normal name=Normal dimension=4 mean=class=Point name=Unnamed dimension=4 values=[50,1,10,5] sigma=class=Point name=Unnamed dimension=4 values=[1,1,1,1] correlationMatrix=class=CorrelationMatrix dimension=4 implementation=class=MatrixImplementation name=Unnamed rows=4 columns=4 values=[1,0,0,0,0,1,0,0,0,0,1,0,0,0,0,1] operator=class=Less name=Unnamed threshold=-3 physicalstartingPoint=class=Point name=Unnamed dimension=4 values=[50,1,10,5] result=class=FORMResult class=AnalyticalResult standardSpaceDesignPoint=class=Point name=Unnamed dimension=0 values=[] physicalSpaceDesignPoint=class=Point name=Unnamed dimension=0 values=[] limitStateVariable=class=RandomVector implementation=class=ThresholdEvent antecedent=class=CompositeRandomVector function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[] evaluationImplementation=class=NoEvaluation name=Unnamed gradientImplementation=class=NoGradient name=Unnamed hessianImplementation=class=NoHessian name=Unnamed antecedent=class=RandomVectorImplementation operator=class=Less name=Unnamed threshold=0 isStandardPointOriginInFailureSpace=false hasoferReliabilityIndex=0 importanceFactors=class=PointWithDescription name=Unnamed dimension=0 description=[] values=[] classicalImportanceFactors=class=PointWithDescription name=Unnamed dimension=0 description=[] values=[] hasoferReliabilityIndexSensitivity=[] eventProbability=0 generalisedReliabilityIndex=0 eventProbabilitySensitivity=[]
+FORM=class=FORM class=Analytical nearestPointAlgorithm=class=OptimizationAlgorithm implementation=class=SQP class=OptimizationAlgorithmImplementation problem=class=OptimizationProblem implementation=class=NearestPointProblem level function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[x1,x2,x3,x4,y0] evaluationImplementation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[x1,x2,x3,x4] outputVariablesNames=[y0] formulas=[x1+2*x2-3*x3+4*x4] gradientImplementation=class=SymbolicGradient name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[x1,x2,x3,x4] outputVariablesNames=[y0] formulas=[x1+2*x2-3*x3+4*x4] hessianImplementation=class=SymbolicHessian name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[x1,x2,x3,x4] outputVariablesNames=[y0] formulas=[x1+2*x2-3*x3+4*x4] level value=3 dimension=4 startingPoint=class=Point name=Unnamed dimension=4 values=[1,1,1,1] maximumIterationNumber=100 maximumEvaluationNumber=100000 maximumAbsoluteError=1e-10 maximumRelativeError=1e-10 maximumResidualError=1e-10 maximumConstraintError=1e-10 verbose=false tau=0.5 omega=0.0001 smooth=1.2 event=class=RandomVector implementation=class=ThresholdEventImplementation antecedent=class=CompositeRandomVector function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[E,F,L,I,y0] evaluationImplementation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] gradientImplementation=class=SymbolicGradient name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] hessianImplementation=class=SymbolicHessian name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] antecedent=class=UsualRandomVector distribution=class=Normal name=Normal dimension=4 mean=class=Point name=Unnamed dimension=4 values=[50,1,10,5] sigma=class=Point name=Unnamed dimension=4 values=[1,1,1,1] correlationMatrix=class=CorrelationMatrix dimension=4 implementation=class=MatrixImplementation name=Unnamed rows=4 columns=4 values=[1,0,0,0,0,1,0,0,0,0,1,0,0,0,0,1] operator=class=Less name=Unnamed threshold=-3 physicalstartingPoint=class=Point name=Unnamed dimension=4 values=[50,1,10,5] result=class=FORMResult class=AnalyticalResult standardSpaceDesignPoint=class=Point name=Unnamed dimension=0 values=[] physicalSpaceDesignPoint=class=Point name=Unnamed dimension=0 values=[] limitStateVariable=class=RandomVector implementation=class=ThresholdEventImplementation antecedent=class=CompositeRandomVector function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[] evaluationImplementation=class=NoEvaluation name=Unnamed gradientImplementation=class=NoGradient name=Unnamed hessianImplementation=class=NoHessian name=Unnamed antecedent=class=RandomVectorImplementation operator=class=Less name=Unnamed threshold=0 isStandardPointOriginInFailureSpace=false hasoferReliabilityIndex=0 importanceFactors=class=PointWithDescription name=Unnamed dimension=0 description=[] values=[] classicalImportanceFactors=class=PointWithDescription name=Unnamed dimension=0 description=[] values=[] hasoferReliabilityIndexSensitivity=[] eventProbability=0 generalisedReliabilityIndex=0 eventProbabilitySensitivity=[]
 event probability=0.156426
 generalized reliability index=1.0093
 standard space design point=[-0.03104,0.84194,0.44540,-0.33224]
