+++ lib/test/t_FORM_draw.expout
@@ -1,5 +1,5 @@
 formulas=[-(6+x0^2-x1)]
 myCobyla=class=Cobyla class=OptimizationAlgorithmImplementation problem=class=OptimizationProblem implementation=class=OptimizationProblemImplementation objective=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[] evaluationImplementation=class=NoEvaluation name=Unnamed gradientImplementation=class=NoGradient name=Unnamed hessianImplementation=class=NoHessian name=Unnamed equality constraint=none inequality constraint=none bounds=none minimization=true dimension=0 startingPoint=class=Point name=Unnamed dimension=0 values=[] maximumIterationNumber=100 maximumEvaluationNumber=200 maximumAbsoluteError=1e-10 maximumRelativeError=1e-10 maximumResidualError=1e-10 maximumConstraintError=1e-10 verbose=false rhoBeg=0.1
-FORM=class=FORM class=Analytical nearestPointAlgorithm=class=OptimizationAlgorithm implementation=class=Cobyla class=OptimizationAlgorithmImplementation problem=class=OptimizationProblem implementation=class=OptimizationProblemImplementation objective=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[] evaluationImplementation=class=NoEvaluation name=Unnamed gradientImplementation=class=NoGradient name=Unnamed hessianImplementation=class=NoHessian name=Unnamed equality constraint=none inequality constraint=none bounds=none minimization=true dimension=0 startingPoint=class=Point name=Unnamed dimension=0 values=[] maximumIterationNumber=100 maximumEvaluationNumber=200 maximumAbsoluteError=1e-10 maximumRelativeError=1e-10 maximumResidualError=1e-10 maximumConstraintError=1e-10 verbose=false rhoBeg=0.1 event=class=RandomVector implementation=class=ThresholdEvent antecedent=class=CompositeRandomVector function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[x0,x1,y0] evaluationImplementation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[x0,x1] outputVariablesNames=[y0] formulas=[-(6+x0^2-x1)] gradientImplementation=class=SymbolicGradient name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[x0,x1] outputVariablesNames=[y0] formulas=[-(6+x0^2-x1)] hessianImplementation=class=SymbolicHessian name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[x0,x1] outputVariablesNames=[y0] formulas=[-(6+x0^2-x1)] antecedent=class=UsualRandomVector distribution=class=Normal name=Normal dimension=2 mean=class=Point name=Unnamed dimension=2 values=[5,2.1] sigma=class=Point name=Unnamed dimension=2 values=[3.3,3] correlationMatrix=class=CorrelationMatrix dimension=2 implementation=class=MatrixImplementation name=Unnamed rows=2 columns=2 values=[1,0,0,1] operator=class=Greater name=Unnamed threshold=0 physicalstartingPoint=class=Point name=Unnamed dimension=2 values=[5,2.1] result=class=FORMResult class=AnalyticalResult standardSpaceDesignPoint=class=Point name=Unnamed dimension=0 values=[] physicalSpaceDesignPoint=class=Point name=Unnamed dimension=0 values=[] limitStateVariable=class=RandomVector implementation=class=ThresholdEvent antecedent=class=CompositeRandomVector function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[] evaluationImplementation=class=NoEvaluation name=Unnamed gradientImplementation=class=NoGradient name=Unnamed hessianImplementation=class=NoHessian name=Unnamed antecedent=class=RandomVectorImplementation operator=class=Less name=Unnamed threshold=0 isStandardPointOriginInFailureSpace=false hasoferReliabilityIndex=0 importanceFactors=class=PointWithDescription name=Unnamed dimension=0 description=[] values=[] classicalImportanceFactors=class=PointWithDescription name=Unnamed dimension=0 description=[] values=[] hasoferReliabilityIndexSensitivity=[] eventProbability=0 generalisedReliabilityIndex=0 eventProbabilitySensitivity=[]
+FORM=class=FORM class=Analytical nearestPointAlgorithm=class=OptimizationAlgorithm implementation=class=Cobyla class=OptimizationAlgorithmImplementation problem=class=OptimizationProblem implementation=class=OptimizationProblemImplementation objective=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[] evaluationImplementation=class=NoEvaluation name=Unnamed gradientImplementation=class=NoGradient name=Unnamed hessianImplementation=class=NoHessian name=Unnamed equality constraint=none inequality constraint=none bounds=none minimization=true dimension=0 startingPoint=class=Point name=Unnamed dimension=0 values=[] maximumIterationNumber=100 maximumEvaluationNumber=200 maximumAbsoluteError=1e-10 maximumRelativeError=1e-10 maximumResidualError=1e-10 maximumConstraintError=1e-10 verbose=false rhoBeg=0.1 event=class=RandomVector implementation=class=ThresholdEventImplementation antecedent=class=CompositeRandomVector function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[x0,x1,y0] evaluationImplementation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[x0,x1] outputVariablesNames=[y0] formulas=[-(6+x0^2-x1)] gradientImplementation=class=SymbolicGradient name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[x0,x1] outputVariablesNames=[y0] formulas=[-(6+x0^2-x1)] hessianImplementation=class=SymbolicHessian name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[x0,x1] outputVariablesNames=[y0] formulas=[-(6+x0^2-x1)] antecedent=class=UsualRandomVector distribution=class=Normal name=Normal dimension=2 mean=class=Point name=Unnamed dimension=2 values=[5,2.1] sigma=class=Point name=Unnamed dimension=2 values=[3.3,3] correlationMatrix=class=CorrelationMatrix dimension=2 implementation=class=MatrixImplementation name=Unnamed rows=2 columns=2 values=[1,0,0,1] operator=class=Greater name=Unnamed threshold=0 physicalstartingPoint=class=Point name=Unnamed dimension=2 values=[5,2.1] result=class=FORMResult class=AnalyticalResult standardSpaceDesignPoint=class=Point name=Unnamed dimension=0 values=[] physicalSpaceDesignPoint=class=Point name=Unnamed dimension=0 values=[] limitStateVariable=class=RandomVector implementation=class=ThresholdEventImplementation antecedent=class=CompositeRandomVector function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[] evaluationImplementation=class=NoEvaluation name=Unnamed gradientImplementation=class=NoGradient name=Unnamed hessianImplementation=class=NoHessian name=Unnamed antecedent=class=RandomVectorImplementation operator=class=Less name=Unnamed threshold=0 isStandardPointOriginInFailureSpace=false hasoferReliabilityIndex=0 importanceFactors=class=PointWithDescription name=Unnamed dimension=0 description=[] values=[] classicalImportanceFactors=class=PointWithDescription name=Unnamed dimension=0 description=[] values=[] hasoferReliabilityIndexSensitivity=[] eventProbability=0 generalisedReliabilityIndex=0 eventProbabilitySensitivity=[]
 importance factors=[0.50259,0.49741]
 importance factors (classical)=[0.50259,0.49741]
