+++ lib/test/t_PostAnalyticalControlledImportanceSampling_std.expout
@@ -1,3 +1,3 @@
 FORM probability=0.156427
-PostAnalyticalControlledImportanceSampling=class=PostAnalyticalControlledImportanceSampling derived from class=PostAnalyticalSimulation analyticalResult=class=AnalyticalResult standardSpaceDesignPoint=class=Point name=Standard Space Design Point dimension=4 values=[-0.0310309,0.841941,0.445397,-0.332234] physicalSpaceDesignPoint=class=Point name=Physical Space Design Point dimension=4 values=[49.969,1.84194,10.4454,4.66777] limitStateVariable=class=RandomVector implementation=class=ThresholdEvent antecedent=class=CompositeRandomVector function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[E,F,L,I,y0] evaluationImplementation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] gradientImplementation=class=SymbolicGradient name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] hessianImplementation=class=SymbolicHessian name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] antecedent=class=UsualRandomVector distribution=class=Normal name=Normal dimension=4 mean=class=Point name=Unnamed dimension=4 values=[50,1,10,5] sigma=class=Point name=Unnamed dimension=4 values=[1,1,1,1] correlationMatrix=class=CorrelationMatrix dimension=4 implementation=class=MatrixImplementation name=Unnamed rows=4 columns=4 values=[1,0,0,0,0,1,0,0,0,0,1,0,0,0,0,1] operator=class=Less name=Unnamed threshold=-3 isStandardPointOriginInFailureSpace=false hasoferReliabilityIndex=1.00925 importanceFactors=class=PointWithDescription name=Unnamed dimension=4 description=[,,,] values=[0,0,0,0] classicalImportanceFactors=class=PointWithDescription name=Unnamed dimension=4 description=[,,,] values=[0,0,0,0] hasoferReliabilityIndexSensitivity=[] controlProbability_=0.156427
+PostAnalyticalControlledImportanceSampling=class=PostAnalyticalControlledImportanceSampling derived from class=PostAnalyticalSimulation analyticalResult=class=AnalyticalResult standardSpaceDesignPoint=class=Point name=Standard Space Design Point dimension=4 values=[-0.0310309,0.841941,0.445397,-0.332234] physicalSpaceDesignPoint=class=Point name=Physical Space Design Point dimension=4 values=[49.969,1.84194,10.4454,4.66777] limitStateVariable=class=RandomVector implementation=class=ThresholdEventImplementation antecedent=class=CompositeRandomVector function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[E,F,L,I,y0] evaluationImplementation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] gradientImplementation=class=SymbolicGradient name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] hessianImplementation=class=SymbolicHessian name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] antecedent=class=UsualRandomVector distribution=class=Normal name=Normal dimension=4 mean=class=Point name=Unnamed dimension=4 values=[50,1,10,5] sigma=class=Point name=Unnamed dimension=4 values=[1,1,1,1] correlationMatrix=class=CorrelationMatrix dimension=4 implementation=class=MatrixImplementation name=Unnamed rows=4 columns=4 values=[1,0,0,0,0,1,0,0,0,0,1,0,0,0,0,1] operator=class=Less name=Unnamed threshold=-3 isStandardPointOriginInFailureSpace=false hasoferReliabilityIndex=1.00925 importanceFactors=class=PointWithDescription name=Unnamed dimension=4 description=[,,,] values=[0,0,0,0] classicalImportanceFactors=class=PointWithDescription name=Unnamed dimension=4 description=[,,,] values=[0,0,0,0] hasoferReliabilityIndexSensitivity=[] controlProbability_=0.156427
 PostAnalyticalControlledImportanceSampling result=probabilityEstimate=1.109273e-01 varianceEstimate=7.444407e-04 standard deviation=2.73e-02 coefficient of variation=2.46e-01 confidenceLength(0.95)=1.07e-01 outerSampling=250 blockSize=4
