+++ lib/test/t_StandardEvent_std.expout
@@ -1,10 +1,10 @@
-myStandardEvent=class=RandomVector implementation=class=ThresholdEvent antecedent=class=CompositeRandomVector function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[E,F,L,I,y0] evaluationImplementation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] gradientImplementation=class=SymbolicGradient name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] hessianImplementation=class=SymbolicHessian name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] antecedent=class=UsualRandomVector distribution=class=Normal name=Normal dimension=4 mean=class=Point name=Unnamed dimension=4 values=[0,0,0,0] sigma=class=Point name=Unnamed dimension=4 values=[1,1,1,1] correlationMatrix=class=CorrelationMatrix dimension=4 implementation=class=MatrixImplementation name=Unnamed rows=4 columns=4 values=[1,0,0,0,0,1,0,0,0,0,1,0,0,0,0,1] operator=class=Less name=Unnamed threshold=1
+myStandardEvent=class=RandomVector implementation=class=ThresholdEventImplementation antecedent=class=CompositeRandomVector function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[E,F,L,I,y0] evaluationImplementation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] gradientImplementation=class=SymbolicGradient name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] hessianImplementation=class=SymbolicHessian name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] antecedent=class=UsualRandomVector distribution=class=Normal name=Normal dimension=4 mean=class=Point name=Unnamed dimension=4 values=[0,0,0,0] sigma=class=Point name=Unnamed dimension=4 values=[1,1,1,1] correlationMatrix=class=CorrelationMatrix dimension=4 implementation=class=MatrixImplementation name=Unnamed rows=4 columns=4 values=[1,0,0,0,0,1,0,0,0,0,1,0,0,0,0,1] operator=class=Less name=Unnamed threshold=1
 myStandardEvent realization (as a RandomVector)=class=Point name=Unnamed dimension=1 values=[1]
 myStandardEvent antecedent realization (as a RandomVector)=class=Point name=Unnamed dimension=4 values=[-2.18139,0.350042,-0.355007,1.43725]
 myStandardEvent realization=class=Point name=Unnamed dimension=1 values=[1]
 myStandardEvent antecedent realization=class=Point name=Unnamed dimension=4 values=[-2.29006,-1.28289,-1.31178,-0.0907838]
 myStandardEvent antecedent distribution cdf=0.501067
-myStandardEvent sample=class=Sample name=ThresholdEvent sample implementation=class=SampleImplementation name=ThresholdEvent sample size=10 dimension=1 description=[y0] data=[[1],[1],[1],[0],[1],[1],[1],[1],[0],[0]]
+myStandardEvent sample=class=Sample name=ThresholdEventImplementation sample implementation=class=SampleImplementation name=ThresholdEventImplementation sample size=10 dimension=1 description=[y0] data=[[1],[1],[1],[0],[1],[1],[1],[1],[0],[0]]
 sample mean=class=Point name=Unnamed dimension=4 values=[0.0227422,-0.024877,-0.026565,-0.0219003]
 sample covariance=class=CovarianceMatrix dimension=4 implementation=class=MatrixImplementation name=Unnamed rows=4 columns=4 values=[0.986028,-0.018174,-0.0104775,0.00791811,-0.018174,1.01902,-0.0103922,0.00475431,-0.0104775,-0.0103922,0.974754,-0.0277212,0.00791811,0.00475431,-0.0277212,1.04267]
 Failure probability (Event)=class=Point name=Unnamed dimension=1 values=[0.864]
