+++ python/test/t_AnalyticalResult_std.expout
@@ -1,4 +1,4 @@
-result= class=AnalyticalResult standardSpaceDesignPoint=class=Point name=Unnamed dimension=4 values=[1,1,1,1] physicalSpaceDesignPoint=class=Point name=Physical Space Design Point dimension=4 values=[51,2,11,6] limitStateVariable=class=RandomVector implementation=class=ThresholdEvent antecedent=class=CompositeRandomVector function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[E,F,L,I,y0] evaluationImplementation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] gradientImplementation=class=SymbolicGradient name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] hessianImplementation=class=SymbolicHessian name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] antecedent=class=UsualRandomVector distribution=class=Normal name=Normal dimension=4 mean=class=Point name=Unnamed dimension=4 values=[50,1,10,5] sigma=class=Point name=Unnamed dimension=4 values=[1,1,1,1] correlationMatrix=class=CorrelationMatrix dimension=4 implementation=class=MatrixImplementation name=Unnamed rows=4 columns=4 values=[1,0,0,0,0,1,0,0,0,0,1,0,0,0,0,1] operator=class=Less name=Unnamed threshold=-3 isStandardPointOriginInFailureSpace=false hasoferReliabilityIndex=2 importanceFactors=class=PointWithDescription name=Unnamed dimension=4 description=[,,,] values=[0,0,0,0] classicalImportanceFactors=class=PointWithDescription name=Unnamed dimension=4 description=[,,,] values=[0,0,0,0] hasoferReliabilityIndexSensitivity=[]
+result= class=AnalyticalResult standardSpaceDesignPoint=class=Point name=Unnamed dimension=4 values=[1,1,1,1] physicalSpaceDesignPoint=class=Point name=Physical Space Design Point dimension=4 values=[51,2,11,6] limitStateVariable=class=RandomVector implementation=class=ThresholdEventImplementation antecedent=class=CompositeRandomVector function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[E,F,L,I,y0] evaluationImplementation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] gradientImplementation=class=SymbolicGradient name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] hessianImplementation=class=SymbolicHessian name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] antecedent=class=UsualRandomVector distribution=class=Normal name=Normal dimension=4 mean=class=Point name=Unnamed dimension=4 values=[50,1,10,5] sigma=class=Point name=Unnamed dimension=4 values=[1,1,1,1] correlationMatrix=class=CorrelationMatrix dimension=4 implementation=class=MatrixImplementation name=Unnamed rows=4 columns=4 values=[1,0,0,0,0,1,0,0,0,0,1,0,0,0,0,1] operator=class=Less name=Unnamed threshold=-3 isStandardPointOriginInFailureSpace=false hasoferReliabilityIndex=2 importanceFactors=class=PointWithDescription name=Unnamed dimension=4 description=[,,,] values=[0,0,0,0] classicalImportanceFactors=class=PointWithDescription name=Unnamed dimension=4 description=[,,,] values=[0,0,0,0] hasoferReliabilityIndexSensitivity=[]
 standard space design point= [1.000000,1.000000,1.000000,1.000000]
 physical space design point= [51.000000,2.000000,11.000000,6.000000]
 is standard point origin in failure space?  False
