+++ python/test/t_MonteCarlo_std.expout
@@ -1,4 +1,4 @@
-MonteCarlo= class=MonteCarlo derived from class=EventSimulation event=class=Event name=Unnamed implementation=class=ThresholdEvent antecedent=class=CompositeRandomVector function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[E,F,L,I,d] evaluationImplementation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[d] formulas=[-F*L^3/(3*E*I)] gradientImplementation=class=SymbolicGradient name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[d] formulas=[-F*L^3/(3*E*I)] hessianImplementation=class=SymbolicHessian name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[d] formulas=[-F*L^3/(3*E*I)] antecedent=class=UsualRandomVector distribution=class=Normal name=Normal dimension=4 mean=class=Point name=Unnamed dimension=4 values=[50,1,10,5] sigma=class=Point name=Unnamed dimension=4 values=[1,1,1,1] correlationMatrix=class=CorrelationMatrix dimension=4 implementation=class=MatrixImplementation name=Unnamed rows=4 columns=4 values=[1,0,0,0,0,1,0,0,0,0,1,0,0,0,0,1] operator=class=Less name=Unnamed threshold=-3 maximumOuterSampling=250 maximumCoefficientOfVariation=0.1 maximumStandardDeviation=0 blockSize=4
+MonteCarlo= class=MonteCarlo derived from class=EventSimulation event=class=Event name=Unnamed implementation=class=ThresholdEventImplementation antecedent=class=CompositeRandomVector function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[E,F,L,I,d] evaluationImplementation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[d] formulas=[-F*L^3/(3*E*I)] gradientImplementation=class=SymbolicGradient name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[d] formulas=[-F*L^3/(3*E*I)] hessianImplementation=class=SymbolicHessian name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[d] formulas=[-F*L^3/(3*E*I)] antecedent=class=UsualRandomVector distribution=class=Normal name=Normal dimension=4 mean=class=Point name=Unnamed dimension=4 values=[50,1,10,5] sigma=class=Point name=Unnamed dimension=4 values=[1,1,1,1] correlationMatrix=class=CorrelationMatrix dimension=4 implementation=class=MatrixImplementation name=Unnamed rows=4 columns=4 values=[1,0,0,0,0,1,0,0,0,0,1,0,0,0,0,1] operator=class=Less name=Unnamed threshold=-3 maximumOuterSampling=250 maximumCoefficientOfVariation=0.1 maximumStandardDeviation=0 blockSize=4
 MonteCarlo result= probabilityEstimate=1.465054e-01 varianceEstimate=2.109726e-04 standard deviation=1.45e-02 coefficient of variation=9.91e-02 confidenceLength(0.95)=5.69e-02 outerSampling=186 blockSize=4
-MonteCarlo= class=MonteCarlo derived from class=EventSimulation event=class=Event name=Unnamed implementation=class=ThresholdEvent antecedent=class=CompositeRandomVector function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[E,F,L,I,d] evaluationImplementation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[d] formulas=[-F*L^3/(3*E*I)] gradientImplementation=class=SymbolicGradient name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[d] formulas=[-F*L^3/(3*E*I)] hessianImplementation=class=SymbolicHessian name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[d] formulas=[-F*L^3/(3*E*I)] antecedent=class=UsualRandomVector distribution=class=Normal name=Normal dimension=4 mean=class=Point name=Unnamed dimension=4 values=[50,1,10,5] sigma=class=Point name=Unnamed dimension=4 values=[1,1,1,1] correlationMatrix=class=CorrelationMatrix dimension=4 implementation=class=MatrixImplementation name=Unnamed rows=4 columns=4 values=[1,0,0,0,0,1,0,0,0,0,1,0,0,0,0,1] operator=class=Less name=Unnamed threshold=-3 maximumOuterSampling=250 maximumCoefficientOfVariation=0 maximumStandardDeviation=0.1 blockSize=4
+MonteCarlo= class=MonteCarlo derived from class=EventSimulation event=class=Event name=Unnamed implementation=class=ThresholdEventImplementation antecedent=class=CompositeRandomVector function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[E,F,L,I,d] evaluationImplementation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[d] formulas=[-F*L^3/(3*E*I)] gradientImplementation=class=SymbolicGradient name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[d] formulas=[-F*L^3/(3*E*I)] hessianImplementation=class=SymbolicHessian name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[d] formulas=[-F*L^3/(3*E*I)] antecedent=class=UsualRandomVector distribution=class=Normal name=Normal dimension=4 mean=class=Point name=Unnamed dimension=4 values=[50,1,10,5] sigma=class=Point name=Unnamed dimension=4 values=[1,1,1,1] correlationMatrix=class=CorrelationMatrix dimension=4 implementation=class=MatrixImplementation name=Unnamed rows=4 columns=4 values=[1,0,0,0,0,1,0,0,0,0,1,0,0,0,0,1] operator=class=Less name=Unnamed threshold=-3 maximumOuterSampling=250 maximumCoefficientOfVariation=0 maximumStandardDeviation=0.1 blockSize=4
 MonteCarlo result= probabilityEstimate=1.500000e-01 varianceEstimate=8.250000e-03 standard deviation=9.08e-02 coefficient of variation=6.06e-01 confidenceLength(0.95)=3.56e-01 outerSampling=5 blockSize=4
