+++ python/test/t_ProbabilitySimulationAlgorithm_draw.expout
@@ -1,2 +1,2 @@
-MonteCarlo= class=ProbabilitySimulationAlgorithm experiment=class=MonteCarloExperiment name=Unnamed distribution=class=Normal name=Normal dimension=4 mean=class=Point name=Unnamed dimension=4 values=[50,1,10,5] sigma=class=Point name=Unnamed dimension=4 values=[1,1,1,1] correlationMatrix=class=CorrelationMatrix dimension=4 implementation=class=MatrixImplementation name=Unnamed rows=4 columns=4 values=[1,0,0,0,0,1,0,0,0,0,1,0,0,0,0,1] size=10 derived from class=EventSimulation event=class=RandomVector implementation=class=ThresholdEvent antecedent=class=CompositeRandomVector function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[E,F,L,I,y0] evaluationImplementation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3.*E*I)] gradientImplementation=class=SymbolicGradient name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3.*E*I)] hessianImplementation=class=SymbolicHessian name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3.*E*I)] antecedent=class=UsualRandomVector distribution=class=Normal name=Normal dimension=4 mean=class=Point name=Unnamed dimension=4 values=[50,1,10,5] sigma=class=Point name=Unnamed dimension=4 values=[1,1,1,1] correlationMatrix=class=CorrelationMatrix dimension=4 implementation=class=MatrixImplementation name=Unnamed rows=4 columns=4 values=[1,0,0,0,0,1,0,0,0,0,1,0,0,0,0,1] operator=class=Less name=Unnamed threshold=-3 maximumOuterSampling=500 maximumCoefficientOfVariation=0.05 maximumStandardDeviation=0 blockSize=10
+MonteCarlo= class=ProbabilitySimulationAlgorithm experiment=class=MonteCarloExperiment name=Unnamed distribution=class=Normal name=Normal dimension=4 mean=class=Point name=Unnamed dimension=4 values=[50,1,10,5] sigma=class=Point name=Unnamed dimension=4 values=[1,1,1,1] correlationMatrix=class=CorrelationMatrix dimension=4 implementation=class=MatrixImplementation name=Unnamed rows=4 columns=4 values=[1,0,0,0,0,1,0,0,0,0,1,0,0,0,0,1] size=10 derived from class=EventSimulation event=class=RandomVector implementation=class=ThresholdEventImplementation antecedent=class=CompositeRandomVector function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[E,F,L,I,y0] evaluationImplementation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3.*E*I)] gradientImplementation=class=SymbolicGradient name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3.*E*I)] hessianImplementation=class=SymbolicHessian name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3.*E*I)] antecedent=class=UsualRandomVector distribution=class=Normal name=Normal dimension=4 mean=class=Point name=Unnamed dimension=4 values=[50,1,10,5] sigma=class=Point name=Unnamed dimension=4 values=[1,1,1,1] correlationMatrix=class=CorrelationMatrix dimension=4 implementation=class=MatrixImplementation name=Unnamed rows=4 columns=4 values=[1,0,0,0,0,1,0,0,0,0,1,0,0,0,0,1] operator=class=Less name=Unnamed threshold=-3 maximumOuterSampling=500 maximumCoefficientOfVariation=0.05 maximumStandardDeviation=0 blockSize=10
 MonteCarlo result= probabilityEstimate=1.415730e-01 varianceEstimate=5.008983e-05 standard deviation=7.08e-03 coefficient of variation=5.00e-02 confidenceLength(0.95)=2.77e-02 outerSampling=267 blockSize=10
