+++ python/test/t_StandardEvent_std.expout
@@ -1,7 +1,7 @@
-myStandardEvent= class=RandomVector implementation=class=ThresholdEvent antecedent=class=CompositeRandomVector function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[E,F,L,I,y0] evaluationImplementation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] gradientImplementation=class=SymbolicGradient name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] hessianImplementation=class=SymbolicHessian name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] antecedent=class=UsualRandomVector distribution=class=Normal name=Normal dimension=4 mean=class=Point name=Unnamed dimension=4 values=[0,0,0,0] sigma=class=Point name=Unnamed dimension=4 values=[1,1,1,1] correlationMatrix=class=CorrelationMatrix dimension=4 implementation=class=MatrixImplementation name=Unnamed rows=4 columns=4 values=[1,0,0,0,0,1,0,0,0,0,1,0,0,0,0,1] operator=class=Less name=Unnamed threshold=1
+myStandardEvent= class=RandomVector implementation=class=ThresholdEventImplementation antecedent=class=CompositeRandomVector function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[E,F,L,I,y0] evaluationImplementation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] gradientImplementation=class=SymbolicGradient name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] hessianImplementation=class=SymbolicHessian name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[E,F,L,I] outputVariablesNames=[y0] formulas=[-F*L^3/(3*E*I)] antecedent=class=UsualRandomVector distribution=class=Normal name=Normal dimension=4 mean=class=Point name=Unnamed dimension=4 values=[0,0,0,0] sigma=class=Point name=Unnamed dimension=4 values=[1,1,1,1] correlationMatrix=class=CorrelationMatrix dimension=4 implementation=class=MatrixImplementation name=Unnamed rows=4 columns=4 values=[1,0,0,0,0,1,0,0,0,0,1,0,0,0,0,1] operator=class=Less name=Unnamed threshold=1
 myStandardEvent (as a RandomVector) realization = class=Point name=Unnamed dimension=1 values=[1]
 myStandardEvent realization= class=Point name=Unnamed dimension=1 values=[1]
-myStandardEvent sample= class=Sample name=ThresholdEvent sample implementation=class=SampleImplementation name=ThresholdEvent sample size=10 dimension=1 description=[y0] data=[[1],[1],[1],[1],[1],[0],[1],[1],[1],[1]]
+myStandardEvent sample= class=Sample name=ThresholdEventImplementation sample implementation=class=SampleImplementation name=ThresholdEventImplementation sample size=10 dimension=1 description=[y0] data=[[1],[1],[1],[1],[1],[0],[1],[1],[1],[1]]
 Failure probability (Event)=0.873500
 Failure probability (StandardEvent)=0.864500
 Error trying to build myStandardEvent3
