+++ python/test/t_StrongMaximumTest_std.expout
@@ -1,4 +1,4 @@
-myTest= class=StrongMaximumTest event=class=RandomVector implementation=class=ThresholdEvent antecedent=class=CompositeRandomVector function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[x1,x2,y0] evaluationImplementation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[x1,x2] outputVariablesNames=[y0] formulas=[(x1-0.3)*(x1-0.3)-x2*x2] gradientImplementation=class=SymbolicGradient name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[x1,x2] outputVariablesNames=[y0] formulas=[(x1-0.3)*(x1-0.3)-x2*x2] hessianImplementation=class=SymbolicHessian name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[x1,x2] outputVariablesNames=[y0] formulas=[(x1-0.3)*(x1-0.3)-x2*x2] antecedent=class=UsualRandomVector distribution=class=Normal name=Normal dimension=2 mean=class=Point name=Unnamed dimension=2 values=[0,0] sigma=class=Point name=Unnamed dimension=2 values=[1,1] correlationMatrix=class=CorrelationMatrix dimension=2 implementation=class=MatrixImplementation name=Unnamed rows=2 columns=2 values=[1,0,0,1] operator=class=Greater name=Unnamed threshold=10 standardSpaceDesignPoint=class=Point name=Unnamed dimension=2 values=[-2.86228,0] importanceLevel=0.01 accuracyLevel=2 confidenceLevel=0.999999 designPointVicinity=0.522209 pointNumber=54 deltaEpsilon=0.457471 nearDesignPointVerifyingEventPoints=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=0 dimension=2 data=[] nearDesignPointVerifyingEventValues=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=0 dimension=1 data=[] farDesignPointVerifyingEventPoints=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=0 dimension=2 data=[] farDesignPointVerifyingEventValues=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=0 dimension=1 data=[] nearDesignPointViolatingEventPoints=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=0 dimension=2 data=[] nearDesignPointViolatingEventValues=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=0 dimension=1 data=[] farDesignPointViolatingEventPoints=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=0 dimension=2 data=[] farDesignPointViolatingEventValues=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=0 dimension=1 data=[]
+myTest= class=StrongMaximumTest event=class=RandomVector implementation=class=ThresholdEventImplementation antecedent=class=CompositeRandomVector function=class=Function name=Unnamed implementation=class=FunctionImplementation name=Unnamed description=[x1,x2,y0] evaluationImplementation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[x1,x2] outputVariablesNames=[y0] formulas=[(x1-0.3)*(x1-0.3)-x2*x2] gradientImplementation=class=SymbolicGradient name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[x1,x2] outputVariablesNames=[y0] formulas=[(x1-0.3)*(x1-0.3)-x2*x2] hessianImplementation=class=SymbolicHessian name=Unnamed evaluation=class=SymbolicEvaluation name=Unnamed inputVariablesNames=[x1,x2] outputVariablesNames=[y0] formulas=[(x1-0.3)*(x1-0.3)-x2*x2] antecedent=class=UsualRandomVector distribution=class=Normal name=Normal dimension=2 mean=class=Point name=Unnamed dimension=2 values=[0,0] sigma=class=Point name=Unnamed dimension=2 values=[1,1] correlationMatrix=class=CorrelationMatrix dimension=2 implementation=class=MatrixImplementation name=Unnamed rows=2 columns=2 values=[1,0,0,1] operator=class=Greater name=Unnamed threshold=10 standardSpaceDesignPoint=class=Point name=Unnamed dimension=2 values=[-2.86228,0] importanceLevel=0.01 accuracyLevel=2 confidenceLevel=0.999999 designPointVicinity=0.522209 pointNumber=54 deltaEpsilon=0.457471 nearDesignPointVerifyingEventPoints=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=0 dimension=2 data=[] nearDesignPointVerifyingEventValues=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=0 dimension=1 data=[] farDesignPointVerifyingEventPoints=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=0 dimension=2 data=[] farDesignPointVerifyingEventValues=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=0 dimension=1 data=[] nearDesignPointViolatingEventPoints=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=0 dimension=2 data=[] nearDesignPointViolatingEventValues=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=0 dimension=1 data=[] farDesignPointViolatingEventPoints=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=0 dimension=2 data=[] farDesignPointViolatingEventValues=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=0 dimension=1 data=[]
 Beta = 2.862278
 Discretised sphere radius = 5.481097
 PointNumber =  54
